split installation test out from where we use py310 syntax
