+++ tests/test_dependencies.py
@@ -111,17 +111,3 @@ def test_has_at_most_one_pinned_dependency(
     assert (
         len(filtered_requirements) == int(expect_pin)
     ), f"{log_msg}.\n{pprint.pformat(environment)}"
-
-
-def test_valid_numpy_is_installed(cfg_requirements):
-    filtered_requirements = []
-    for req in cfg_requirements:
-        if req.marker.evaluate():
-            filtered_requirements.append(req)
-
-    assert (len(filtered_requirements) == 1), "Expected exactly one pin."
-
-    item, = filtered_requirements[0].specifier
-
-    import numpy
-    assert item.version == numpy.__version__
